requirements.txt
crispy-bootstrap4==2024.1
Django==5.0.7
django-crispy-forms==2.2
django-environ==0.11.2
django-filter==24.2
django-localflavor==4.0
gunicorn==22.0.0
pillow==10.4.0
psycopg2-binary==2.9.9